Adds forge git repo as extra dep
So that this can be compiled automatically without having forge
cloned on the same machine in a pre-defined space. Was useful for
initial testing but this will make it easier for anybody else to
compile the project on their machines.

Long term I should probably consider publishing forge on
hackage/stackage. But that's not going to happen any time soon.